phase 2

+++ src/main/java/hhvitek/documentmanager/protocol/ProtocolService.java
@@ -2,20 +2,20 @@ package hhvitek.documentmanager.protocol;
 
 import java.util.List;
 
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import hhvitek.documentmanager.document.Document;
 import hhvitek.documentmanager.document.DocumentRepository;
+import hhvitek.documentmanager.error.ApiException;
 
 @Service
 public class ProtocolService {
-	private final ProtocolValidator protocolValidator;
 	private final ProtocolRepository protocolRepository;
 	private final DocumentRepository documentRepository;
 
-	public ProtocolService(ProtocolValidator protocolValidator, ProtocolRepository protocolRepository, DocumentRepository documentRepository) {
-		this.protocolValidator = protocolValidator;
+	public ProtocolService(ProtocolRepository protocolRepository, DocumentRepository documentRepository) {
 		this.protocolRepository = protocolRepository;
 		this.documentRepository = documentRepository;
 	}
@@ -26,44 +26,56 @@ public class ProtocolService {
 	}
 
 	public Protocol getOne(Integer id) {
-		return protocolRepository.findById(id).orElseThrow();
+		return protocolRepository.findById(id).orElseThrow(() -> new ApiException("No such protocol found with id=" + id));
 	}
 
 	@Transactional
-	public Protocol create(ProtocolDTO protocol) {
-		protocolValidator.validateDocumentIds(protocol); // all documents exists now
-		List<Document> documents = documentRepository.findAllById(protocol.getDocumentIds());
+	public Protocol create(ProtocolDTO dto) {
+		Protocol newProtocol = dto.toProtocolWithoutDocuments();
 
-		Protocol newProtocol = protocol.toProtocol();
-		newProtocol.setDocuments(documents);
+		if (CollectionUtils.isEmpty(dto.getDocumentIds())) {
+			throw new ApiException("Protocol must contain at least one document!");
+		}
+
+		setDocumentsIfNotFoundThrow(newProtocol, dto);
 
 		return protocolRepository.save(newProtocol);
 	}
 
+	private void setDocumentsIfNotFoundThrow(Protocol protocol, ProtocolDTO dto) {
+		List<Document> documents = documentRepository.findAllById(dto.getDocumentIds());
+		if (documents.size() != dto.getDocumentIds().size()) {
+			throw new ApiException("Invalid documents. At least one document was not found.");
+		}
+		protocol.setDocuments(documents);
+	}
+
 	@Transactional
-	public Protocol update(Integer id, ProtocolDTO protocol) {
+	public Protocol update(Integer id, ProtocolDTO dto) {
 		Protocol oldProtocol = getOne(id); // it exists
 
-		if (protocol.hasAnyDocumentIds()) {
-			protocolValidator.validateDocumentIds(protocol); // lets validate protocol ids and set them
-			List<Document> documents = documentRepository.findAllById(protocol.getDocumentIds());
-			oldProtocol.setDocuments(documents);
+		if (dto.hasAnyDocumentIds()) { // if document ids are part of passed DTO, otherwise retain old values
+			setDocumentsIfNotFoundThrow(oldProtocol, dto);
 		}
 
-		protocol.modify(oldProtocol);
+		dto.modify(oldProtocol);
 
 		return protocolRepository.save(oldProtocol);
 	}
 
-	@Transactional
+
 	public Protocol updateState(Integer id, ProtocolState state) {
-		Protocol protocol = getOne(id);
-		protocol.setState(state);
-		return protocolRepository.save(protocol);
+		Protocol oldProtocol = getOne(id);
+		oldProtocol.setState(state);
+		return protocolRepository.save(oldProtocol);
 	}
 
 	public void delete(Integer id) {
-		protocolRepository.deleteById(id);
+		if (protocolRepository.existsById(id)) {
+			protocolRepository.deleteById(id); // may throw EmptyResultDataAccessException if no such id exists
+		} else {
+			throw new ApiException("Cannot delete protocol. No such protocol found with id=" + id);
+		}
 	}
